requirements.txt
appdirs==1.4.3
appnope==0.1.0
attrs==18.2.0
backcall==0.1.0
black==18.9b0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
cssselect==1.0.3
cssutils==1.0.2
decorator==4.3.0
Django==2.1.4
django-environ==0.4.5
django-extensions==2.1.4
django-filter==2.0.0
djangorestframework==3.9.0
emails==0.5.15
greenlet==0.4.15
idna==2.7
ipdb==0.11
ipython==7.2.0
ipython-genutils==0.2.0
jedi==0.13.1
lxml==4.2.5
Markdown==3.0.1
msgpack==0.6.0
neovim==0.3.1
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
premailer==3.2.0
prompt-toolkit==2.0.7
psycopg2==2.7.6.1
ptyprocess==0.6.0
Pygments==2.3.0
pynvim==0.3.1
python-dateutil==2.7.5
pytz==2018.7
requests==2.20.1
six==1.11.0
toml==0.10.0
traitlets==4.3.2
urllib3==1.24.1
wcwidth==0.1.7
